+++ packages/plugins/other/visitor-plugin-common/src/client-side-base-visitor.ts
@@ -1,5 +1,5 @@
 import { basename, extname } from 'path';
-import { oldVisit, Types } from '@graphql-codegen/plugin-helpers';
+import { normalizeImportExtension, oldVisit, Types } from '@graphql-codegen/plugin-helpers';
 import { optimizeDocumentNode } from '@graphql-tools/optimize';
 import autoBind from 'auto-bind';
 import { pascalCase } from 'change-case-all';
@@ -408,7 +408,7 @@ export class ClientSideBaseVisitor<
       }
 
       let metaString = '';
-      if (this._onExecutableDocumentNode && node.kind === Kind.OPERATION_DEFINITION) {
+      if (this._onExecutableDocumentNode) {
         const meta = this._getGraphQLCodegenMetadata(node, definitions);
 
         if (meta) {
@@ -425,7 +425,9 @@ export class ClientSideBaseVisitor<
 
     if (this.config.documentMode === DocumentMode.string) {
       if (node.kind === Kind.FRAGMENT_DEFINITION) {
-        return `new TypedDocumentString(\`${doc}\`, ${JSON.stringify({ fragmentName: node.name.value })})`;
+        const meta = this._getGraphQLCodegenMetadata(node, gqlTag([doc]).definitions);
+
+        return `new TypedDocumentString(\`${doc}\`, ${JSON.stringify({ fragmentName: node.name.value, ...meta })})`;
       }
 
       if (this._onExecutableDocumentNode && node.kind === Kind.OPERATION_DEFINITION) {
@@ -448,15 +450,17 @@ export class ClientSideBaseVisitor<
   }
 
   protected _getGraphQLCodegenMetadata(
-    node: OperationDefinitionNode,
+    node: OperationDefinitionNode | FragmentDefinitionNode,
     definitions?: ReadonlyArray<DefinitionNode>
   ): Record<string, any> | void | undefined {
     let meta: Record<string, any> | void | undefined;
 
-    meta = this._onExecutableDocumentNode({
-      kind: Kind.DOCUMENT,
-      definitions,
-    });
+    if (node.kind === Kind.OPERATION_DEFINITION) {
+      meta = this._onExecutableDocumentNode({
+        kind: Kind.DOCUMENT,
+        definitions,
+      });
+    }
 
     const deferredFields = this._findDeferredFields(node);
     if (Object.keys(deferredFields).length) {
@@ -469,7 +473,9 @@ export class ClientSideBaseVisitor<
     return meta;
   }
 
-  protected _findDeferredFields(node: OperationDefinitionNode): { [fargmentName: string]: string[] } {
+  protected _findDeferredFields(node: OperationDefinitionNode | FragmentDefinitionNode): {
+    [fargmentName: string]: string[];
+  } {
     const deferredFields: { [fargmentName: string]: string[] } = {};
     const queue: SelectionNode[] = [...node.selectionSet.selections];
     while (queue.length) {
@@ -604,7 +610,12 @@ export class ClientSideBaseVisitor<
   private clearExtension(path: string): string {
     const extension = extname(path);
 
-    if (!this.config.emitLegacyCommonJSImports && extension === '.js') {
+    const importExtension = normalizeImportExtension({
+      emitLegacyCommonJSImports: this.config.emitLegacyCommonJSImports,
+      importExtension: this.config.importExtension,
+    });
+
+    if (extension === importExtension) {
       return path;
     }
 
@@ -646,9 +657,10 @@ export class ClientSideBaseVisitor<
         if (this._collectedOperations.length > 0) {
           if (this.config.importDocumentNodeExternallyFrom === 'near-operation-file' && this._documents.length === 1) {
             let documentPath = `./${this.clearExtension(basename(this._documents[0].location))}`;
-            if (!this.config.emitLegacyCommonJSImports) {
-              documentPath += '.js';
-            }
+            documentPath += normalizeImportExtension({
+              emitLegacyCommonJSImports: this.config.emitLegacyCommonJSImports,
+              importExtension: this.config.importExtension,
+            });
 
             this._imports.add(`import * as Operations from '${documentPath}';`);
           } else {
@@ -672,6 +684,10 @@ export class ClientSideBaseVisitor<
       options.excludeFragments || this.config.globalNamespace || this.config.documentMode !== DocumentMode.graphQLTag;
 
     if (!excludeFragments) {
+      const importExtension = normalizeImportExtension({
+        emitLegacyCommonJSImports: this.config.emitLegacyCommonJSImports,
+        importExtension: this.config.importExtension,
+      });
       const deduplicatedImports = Object.values(groupBy(this.config.fragmentImports, fi => fi.importSource.path))
         .map(
           (fragmentImports): ImportDeclaration<FragmentImport> => ({
@@ -684,6 +700,7 @@ export class ClientSideBaseVisitor<
               ),
             },
             emitLegacyCommonJSImports: this.config.emitLegacyCommonJSImports,
+            importExtension,
           })
         )
         .filter(fragmentImport => fragmentImport.outputPath !== fragmentImport.importSource.path);